motor 4 format display set time

+++ MayBom4Step/App/app_pump/lib/motor.dart
@@ -7,7 +7,7 @@ class Motor {
   double pulse;
   double microStep;
   double speed;
-  double setTime;
+  int setTime;
 
   Motor(this.noM, this.pulse, this.microStep, this.speed, this.setTime);
